+++ examples/test_example_context_manager_fail.py
@@ -9,13 +9,20 @@ This test is useful for testing:
 
 from pytest_check import check
 
+
 def test_3_failed_checks():
-    with check: assert 1 == 0
-    with check: assert 1 > 2
-    with check: assert 1 < 5 < 4
+    with check:
+        assert 1 == 0
+    with check:
+        assert 1 > 2
+    with check:
+        assert 1 < 5 < 4
 
 
 def test_messages():
-    with check("first fail"): assert 1 == 0
-    with check("second fail"): assert 1 > 2
-    with check("third fail"): assert 1 < 5 < 4
+    with check("first fail"):
+        assert 1 == 0
+    with check("second fail"):
+        assert 1 > 2
+    with check("third fail"):
+        assert 1 < 5 < 4